Add files via upload
The signup/login form is a sleek and user-friendly web application designed using HTML, CSS, and JavaScript. It incorporates form validation to ensure data integrity. HTML provides the structural framework, defining form fields for users to input their name, username, email, and password. The JavaScript script at the bottom of the HTML includes an event listener for the form submission. When the form is submitted, it prevents the default form submission, displays an alert, and then resets the form fields using the reset() method. JavaScript is utilized for client-side form validation, enhancing the interactivity of the form. The script ensures that users input valid data by checking the name and username for correctness, verifying the email format, and enforcing password complexity requirements. The form also includes links for password recovery and easy navigation between the signup and login pages.

Deployed Link - https://signuploginform-validation.netlify.app/

Note :- Open the form or link in PC or Desktop for better view.